Rebuild the plugin system in a more OO approach.
This rewrites a lot of hook.py, ands moves a bunch of stuff from loader.py to pluginmanager.py.

It also removes @hook.singlethread, in favor of singlethread=True param.

It adds support for multiple commands in a hook "@hook.command(["cmd1", "cmd2"])", as well as multiple regexes and events in one regex/event hook.

It does not yet support reloading plugins, though there are just a few more lines of code to write to fix that.